+++ test/quantization/test_quantize.py
@@ -1,5 +1,4 @@
 import unittest
-import math
 import torch
 import torch.nn as nn
 import torch.nn.quantized as nnq
@@ -8,13 +7,11 @@ import torch.nn.intrinsic.quantized as nniq
 import torch.nn.intrinsic.qat as nniqat
 from torch.nn.utils.rnn import PackedSequence
 from torch.quantization import \
-    get_observer_dict, default_weight_observer, \
+    default_weight_observer, \
     quantize, prepare, convert, prepare_qat, quantize_qat, fuse_modules, \
     quantize_dynamic, default_qconfig, default_debug_qconfig, default_qat_qconfig, \
-    default_dynamic_qconfig, per_channel_dynamic_qconfig, HistogramObserver, MinMaxObserver, \
-    PerChannelMinMaxObserver, RecordingObserver, MovingAverageMinMaxObserver, \
-    MovingAveragePerChannelMinMaxObserver, QuantWrapper, default_eval_fn, \
-    float16_dynamic_qconfig, MinMaxDynamicQuantObserver
+    default_dynamic_qconfig, per_channel_dynamic_qconfig, HistogramObserver, QuantWrapper, default_eval_fn, \
+    float16_dynamic_qconfig
 
 from torch.quantization import QConfig
 from torch.quantization import default_histogram_observer
@@ -1453,338 +1450,6 @@ class TestFusion(QuantizationTestCase):
             self.assertEqual(type(model.bn2), nn.Identity)
 
         checkQAT(model)
-
-class TestObserver(QuantizationTestCase):
-    @given(qdtype=st.sampled_from((torch.qint8, torch.quint8)),
-           qscheme=st.sampled_from((torch.per_tensor_affine, torch.per_tensor_symmetric)),
-           reduce_range=st.booleans())
-    def test_per_tensor_observers(self, qdtype, qscheme, reduce_range):
-        # reduce_range cannot be true for symmetric quantization with uint8
-        if qdtype == torch.quint8 and qscheme == torch.per_tensor_symmetric:
-            reduce_range = False
-        ObserverList = [MinMaxObserver(dtype=qdtype, qscheme=qscheme, reduce_range=reduce_range),
-                        MovingAverageMinMaxObserver(averaging_constant=0.5,
-                                                    dtype=qdtype,
-                                                    qscheme=qscheme,
-                                                    reduce_range=reduce_range)]
-        for myobs in ObserverList:
-            # Calculate Qparams should return with a warning for observers with no data
-            qparams = myobs.calculate_qparams()
-            if type(myobs) == MinMaxObserver:
-                x = torch.tensor([1.0, 2.0, 2.0, 3.0, 4.0, 5.0, 6.0])
-                y = torch.tensor([4.0, 5.0, 5.0, 6.0, 7.0, 8.0])
-            else:
-                # Moving average of min/max for x and y matches that of
-                # extreme values for x/y used for minmax observer
-                x = torch.tensor([0.0, 2.0, 2.0, 3.0, 4.0, 5.0, 6.0])
-                y = torch.tensor([2.0, 5.0, 5.0, 6.0, 7.0, 10.0])
-
-            result = myobs(x)
-            result = myobs(y)
-            self.assertEqual(result, y)
-            self.assertEqual(myobs.min_val, 1.0)
-            self.assertEqual(myobs.max_val, 8.0)
-            qparams = myobs.calculate_qparams()
-            if reduce_range:
-                if qscheme == torch.per_tensor_symmetric:
-                    ref_scale = 0.062745 * 255 / 127
-                    ref_zero_point = 0 if qdtype is torch.qint8 else 128
-                else:
-                    ref_scale = 0.0313725 * 255 / 127
-                    ref_zero_point = -64 if qdtype is torch.qint8 else 0
-            else:
-                if qscheme == torch.per_tensor_symmetric:
-                    ref_scale = 0.062745
-                    ref_zero_point = 0 if qdtype is torch.qint8 else 128
-                else:
-                    ref_scale = 0.0313725
-                    ref_zero_point = -128 if qdtype is torch.qint8 else 0
-            self.assertEqual(qparams[1].item(), ref_zero_point)
-            self.assertAlmostEqual(qparams[0].item(), ref_scale, delta=1e-5)
-            state_dict = myobs.state_dict()
-            b = io.BytesIO()
-            torch.save(state_dict, b)
-            b.seek(0)
-            loaded_dict = torch.load(b)
-            for key in state_dict:
-                self.assertEqual(state_dict[key], loaded_dict[key])
-            loaded_obs = MinMaxObserver(dtype=qdtype, qscheme=qscheme, reduce_range=reduce_range)
-            loaded_obs.load_state_dict(loaded_dict)
-            loaded_qparams = loaded_obs.calculate_qparams()
-            self.assertEqual(myobs.min_val, loaded_obs.min_val)
-            self.assertEqual(myobs.max_val, loaded_obs.max_val)
-            self.assertEqual(myobs.calculate_qparams(), loaded_obs.calculate_qparams())
-
-
-    @given(X=hu.tensor(shapes=hu.array_shapes(min_dims=2, max_dims=4,
-                                              min_side=1, max_side=10),
-                       qparams=hu.qparams()),
-           reduce_range=st.booleans())
-    def test_per_tensor_dynamic_quant_observers(self, X, reduce_range):
-
-        X, (scale, zero_point, torch_type) = X
-        x = torch.from_numpy(X)
-
-        obs = MinMaxDynamicQuantObserver(dtype=torch.quint8, reduce_range=reduce_range)
-
-        result = obs(x)
-        qparams = obs.calculate_qparams()
-        ref = torch._choose_qparams_per_tensor(x, reduce_range)
-
-        self.assertEqual(ref[0], qparams[0])
-        self.assertEqual(ref[1], qparams[1])
-
-    def test_tensor_list_observer(self):
-        from torch.quantization.observer import _MinMaxTensorListObserver
-        x = [torch.tensor([1.0, 2.5, 3.5]),
-             torch.tensor([2.0, 4.5, 3.5]),
-             torch.tensor([4.0, 2.5, 3.5]), ]
-        obs = _MinMaxTensorListObserver()
-        obs(x)
-        qparams = obs.calculate_qparams()
-        ref_min_val = []
-        ref_max_val = []
-        ref_qparams = []
-        for i in x:
-            obs_ref = MinMaxObserver()
-            obs_ref(i)
-            ref_min_val.append(obs_ref.min_val)
-            ref_max_val.append(obs_ref.max_val)
-            ref_qparams.append(obs_ref.calculate_qparams())
-        for i in range(len(x)):
-            self.assertEqual(obs.min_val[i], ref_min_val[i])
-            self.assertEqual(obs.max_val[i], ref_max_val[i])
-            self.assertEqual(qparams[0][i], ref_qparams[i][0])
-            self.assertEqual(qparams[1][i], ref_qparams[i][1])
-
-    @given(qdtype=st.sampled_from((torch.qint8, torch.quint8)),
-           qscheme=st.sampled_from((torch.per_channel_affine, torch.per_channel_symmetric)),
-           ch_axis=st.sampled_from((0, 1, 2, 3)), reduce_range=st.booleans())
-    def test_per_channel_observers(self, qdtype, qscheme, ch_axis, reduce_range):
-        # reduce_range cannot be true for symmetric quantization with uint8
-        if qdtype == torch.quint8 and qscheme == torch.per_channel_symmetric:
-            reduce_range = False
-        ObserverList = [PerChannelMinMaxObserver(reduce_range=reduce_range,
-                                                 ch_axis=ch_axis,
-                                                 dtype=qdtype,
-                                                 qscheme=qscheme),
-                        MovingAveragePerChannelMinMaxObserver(averaging_constant=0.5,
-                                                              reduce_range=reduce_range,
-                                                              ch_axis=ch_axis,
-                                                              dtype=qdtype,
-                                                              qscheme=qscheme)]
-
-        for myobs in ObserverList:
-            # Calculate qparams should work for empty observers
-            qparams = myobs.calculate_qparams()
-            x = torch.tensor(
-                [
-                    [[[1.0, 2.0], [2.0, 2.5]], [[3.0, 4.0], [4.5, 6.0]]],
-                    [[[-4.0, -3.0], [5.0, 5.0]], [[6.0, 3.0], [7.0, 8.0]]],
-                ]
-            )
-            if type(myobs) == MovingAveragePerChannelMinMaxObserver:
-                # Scaling the input tensor to model change in min/max values
-                # across batches
-                result = myobs(0.5 * x)
-                result = myobs(1.5 * x)
-                self.assertEqual(result, 1.5 * x)
-            else:
-                result = myobs(x)
-                self.assertEqual(result, x)
-
-            qparams = myobs.calculate_qparams()
-            ref_min_vals = [[1.0, -4.0], [-4.0, 3.0], [-4.0, 2.0], [-4.0, -3.0]]
-            ref_max_vals = [[6.0, 8.0], [5.0, 8.0], [6.0, 8.0], [7.0, 8.0]]
-            per_channel_symmetric_ref_scales = [
-                [0.04705882, 0.06274509],
-                [0.03921569, 0.0627451],
-                [0.04705882, 0.0627451],
-                [0.05490196, 0.0627451],
-            ]
-            per_channel_affine_ref_scales = [
-                [0.02352941, 0.04705882],
-                [0.03529412, 0.03137255],
-                [0.03921569, 0.03137255],
-                [0.04313726, 0.04313726],
-            ]
-            per_channel_affine_qint8_zp = [
-                [-128, -43],
-                [-15, -128],
-                [-26, -128],
-                [-35, -58],
-            ]
-            per_channel_affine_quint8_zp = [[0, 85], [113, 0], [102, 0], [93, 70]]
-
-            self.assertEqual(myobs.min_vals, ref_min_vals[ch_axis])
-            self.assertEqual(myobs.max_vals, ref_max_vals[ch_axis])
-            if qscheme == torch.per_channel_symmetric:
-                ref_scales = per_channel_symmetric_ref_scales[ch_axis]
-                ref_zero_points = [0, 0] if qdtype is torch.qint8 else [128, 128]
-            else:
-                ref_scales = per_channel_affine_ref_scales[ch_axis]
-                ref_zero_points = (
-                    per_channel_affine_qint8_zp[ch_axis]
-                    if qdtype is torch.qint8
-                    else per_channel_affine_quint8_zp[ch_axis]
-                )
-
-            if reduce_range:
-                ref_scales = [s * 255 / 127 for s in ref_scales]
-                ref_zero_points = [math.floor(z / 2) for z in ref_zero_points]
-
-            self.assertTrue(torch.allclose(qparams[0], torch.tensor(ref_scales, dtype=qparams[0].dtype)))
-            self.assertTrue(torch.allclose(qparams[1], torch.tensor(ref_zero_points, dtype=qparams[1].dtype)))
-
-            # Test for serializability
-            state_dict = myobs.state_dict()
-            b = io.BytesIO()
-            torch.save(state_dict, b)
-            b.seek(0)
-            loaded_dict = torch.load(b)
-            for key in state_dict:
-                self.assertEqual(state_dict[key], loaded_dict[key])
-            loaded_obs = PerChannelMinMaxObserver(reduce_range=reduce_range, ch_axis=ch_axis, dtype=qdtype, qscheme=qscheme)
-            loaded_obs.load_state_dict(loaded_dict)
-            loaded_qparams = loaded_obs.calculate_qparams()
-            self.assertEqual(myobs.min_vals, loaded_obs.min_vals)
-            self.assertEqual(myobs.max_vals, loaded_obs.max_vals)
-            self.assertEqual(myobs.calculate_qparams(), loaded_obs.calculate_qparams())
-
-    def test_observer_scriptable(self):
-        obs_list = [MinMaxObserver(), MovingAverageMinMaxObserver(), MinMaxDynamicQuantObserver()]
-        for obs in obs_list:
-            scripted = torch.jit.script(obs)
-
-            x = torch.rand(3, 4)
-            obs(x)
-            scripted(x)
-            self.assertEqual(obs.calculate_qparams(), scripted.calculate_qparams())
-
-            buf = io.BytesIO()
-            torch.jit.save(scripted, buf)
-            buf.seek(0)
-            loaded = torch.jit.load(buf)
-            self.assertEqual(obs.calculate_qparams(), loaded.calculate_qparams())
-
-        # Check TensorListObserver
-        from torch.quantization.observer import _MinMaxTensorListObserver
-        obs = _MinMaxTensorListObserver()
-        scripted = torch.jit.script(obs)
-        x = [torch.rand(3, 4), torch.rand(4, 5)]
-        obs(x)
-        scripted(x)
-        self.assertEqual(obs.calculate_qparams(), scripted.calculate_qparams())
-
-    def test_no_qconfig_propagation(self):
-        model = ModelWithNoQconfigPropagation()
-        model.qconfig = torch.quantization.default_qconfig
-
-        model = prepare(model)
-        self.assertTrue(hasattr(model.fc1, 'qconfig'),
-                        "QConfig is expected to propagate")
-        self.assertFalse(hasattr(model.no_quant_module, 'qconfig'),
-                         "QConfig is expected to NOT propagate")
-
-
-@unittest.skipUnless('fbgemm' in torch.backends.quantized.supported_engines,
-                     " Quantized operations require FBGEMM. FBGEMM is only optimized for CPUs"
-                     " with instruction set support avx2 or newer.")
-class TestRecordHistogramObserver(QuantizationTestCase):
-    def test_record_observer(self):
-        model = AnnotatedSingleLayerLinearModel()
-        model.qconfig = default_debug_qconfig
-        model = prepare(model)
-        # run the evaluation and dump all tensors
-        test_only_eval_fn(model, self.calib_data)
-        test_only_eval_fn(model, self.calib_data)
-        observer_dict = {}
-        get_observer_dict(model, observer_dict)
-
-        self.assertTrue('fc1.module.activation_post_process' in observer_dict.keys(),
-                        'observer is not recorded in the dict')
-        self.assertEqual(len(observer_dict['fc1.module.activation_post_process'].get_tensor_value()), 2 * len(self.calib_data))
-        self.assertEqual(observer_dict['fc1.module.activation_post_process'].get_tensor_value()[0], model(self.calib_data[0][0]))
-
-    @given(qdtype=st.sampled_from((torch.qint8, torch.quint8)),
-           qscheme=st.sampled_from((torch.per_tensor_affine, torch.per_tensor_symmetric)))
-    def test_observer_scriptable(self, qdtype, qscheme):
-        obs = RecordingObserver(dtype=qdtype, qscheme=qscheme)
-        scripted = torch.jit.script(obs)
-
-        x = torch.rand(3, 4)
-        obs(x)
-        scripted(x)
-        self.assertTrue(torch.equal(obs.get_tensor_value()[0], scripted.get_tensor_value()[0]))
-        buf = io.BytesIO()
-        torch.jit.save(scripted, buf)
-        buf.seek(0)
-        loaded = torch.jit.load(buf)
-        self.assertTrue(torch.equal(obs.get_tensor_value()[0], loaded.get_tensor_value()[0]))
-
-    @given(qdtype=st.sampled_from((torch.qint8, torch.quint8)),
-           qscheme=st.sampled_from((torch.per_tensor_affine, torch.per_tensor_symmetric)),
-           reduce_range=st.booleans())
-    def test_histogram_observer(self, qdtype, qscheme, reduce_range):
-        myobs = HistogramObserver(bins=3, dtype=qdtype, qscheme=qscheme, reduce_range=reduce_range)
-        # Calculate qparams should work for empty observers
-        qparams = myobs.calculate_qparams()
-        x = torch.tensor([2.0, 3.0, 4.0, 5.0], requires_grad=True)
-        y = torch.tensor([5.0, 6.0, 7.0, 8.0])
-        out_x = myobs(x)
-        self.assertTrue(out_x.requires_grad)
-        myobs(y)
-        self.assertEqual(myobs.min_val, 2.0)
-        self.assertEqual(myobs.max_val, 8.0)
-        self.assertEqual(myobs.histogram, [2., 3., 3.])
-
-        qparams = myobs.calculate_qparams()
-
-        if reduce_range:
-            if qscheme == torch.per_tensor_symmetric:
-                ref_scale = 0.0470588 * 255 / 127
-                ref_zero_point = 0 if qdtype is torch.qint8 else 128
-            else:
-                ref_scale = 0.0235294 * 255 / 127
-                ref_zero_point = -64 if qdtype is torch.qint8 else 0
-        else:
-            if qscheme == torch.per_tensor_symmetric:
-                ref_scale = 0.0470588
-                ref_zero_point = 0 if qdtype is torch.qint8 else 128
-            else:
-                ref_scale = 0.0235294
-                ref_zero_point = -128 if qdtype is torch.qint8 else 0
-
-        self.assertEqual(qparams[1].item(), ref_zero_point)
-        self.assertAlmostEqual(qparams[0].item(), ref_scale, delta=1e-5)
-        # Test for serializability
-        state_dict = myobs.state_dict()
-        b = io.BytesIO()
-        torch.save(state_dict, b)
-        b.seek(0)
-        loaded_dict = torch.load(b)
-        for key in state_dict:
-            self.assertEqual(state_dict[key], loaded_dict[key])
-        loaded_obs = HistogramObserver(bins=3, dtype=qdtype, qscheme=qscheme, reduce_range=reduce_range)
-        loaded_obs.load_state_dict(loaded_dict)
-        loaded_qparams = loaded_obs.calculate_qparams()
-        self.assertEqual(myobs.min_val, loaded_obs.min_val)
-        self.assertEqual(myobs.max_val, loaded_obs.max_val)
-        self.assertEqual(myobs.histogram, loaded_obs.histogram)
-        self.assertEqual(myobs.bins, loaded_obs.bins)
-        self.assertEqual(myobs.calculate_qparams(), loaded_obs.calculate_qparams())
-
-    def test_histogram_observer_one_sided(self):
-        myobs = HistogramObserver(bins=8, dtype=torch.quint8, qscheme=torch.per_tensor_affine, reduce_range=True)
-        x = torch.tensor([0.0, 0.3, 1.2, 1.7])
-        y = torch.tensor([0.1, 1.3, 2.0, 2.7])
-        myobs(x)
-        myobs(y)
-        self.assertEqual(myobs.min_val, 0)
-        qparams = myobs.calculate_qparams()
-        self.assertEqual(qparams[1].item(), 0)
-
-
 
 if __name__ == '__main__':
     run_tests()
